Create a script to convert the city-population-file of CBS
Problem: The CBS (Centraal Bureau voor Statistieken) shares information about the number op people that live in every city. The data in this file has a wrong format and has too much data.
Solution: This script reads the data from the CBS-JSON-file and writes the relevant data to a new JSON-file, with the correct syntax.