Determine start- and end Date and fix time zone bug
The site is incosistant for clients for different time zones. Bugs that did accure:
* Time zone negative: All days are shifted one day earlier. 26 Febrari should not exist, but the first day (27 Februari 00:00+0) is shown as 26 Februari.
* Months are displayed wrong because of incosistant use of UTC and not-UTC dates: ex: 28 Februari, 29 Februari, 1 Februari, 2 March.

Also, the earliest and latest date of datasets can be determined automatically now. No need to fill in dates by hand.

When dates are invalid, an alert is shown with the invalid datestring and the data is not loaded.